refactoring singleton logic into metaclasses
